Add repository and service layer

+++ Task2/src/main/java/pl/jkkk/task2/logic/repository/PollutionRepository.java
@@ -3,7 +3,7 @@ package pl.jkkk.task2.logic.repository;
 import org.springframework.data.repository.CrudRepository;
 import pl.jkkk.task2.logic.model.Pollution;
 
-public interface PollutionRepository extends CrudRepository<Pollution, Integer> {
+public interface PollutionRepository extends CrudRepository<Pollution, Long> {
 
 }
